third_party/WebKit/LayoutTests/intersection-observer/display-none.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./resources/intersection-observer-test-utils.js"></script>

<style>
pre, #log {
  position: absolute;
  top: 0;
  left: 200px;
}
#target {
  background-color: green;
  width: 100px;
  height: 100px;
}
</style>

<div id="target"></div>

<script>
var entries = [];

runTestCycle(function() {
  assert_greater_than(window.innerWidth, 200, "Window must be at least 200 pixels wide.");
  assert_greater_than(window.innerHeight, 200, "Window must be at least 200 pixels high.");

  var target = document.getElementById("target");
  var observer = new IntersectionObserver(function(changes) {
    entries = entries.concat(changes)
  });
  observer.observe(target);
  entries = entries.concat(observer.takeRecords());
  assert_equals(entries.length, 0, "No initial notifications.");
  runTestCycle(step0, "Intersecting notification after first rAF.");
}, "IntersectionObserver should send a not-intersecting notification for a target that gets display:none.");

function step0() {
  runTestCycle(step1, "Not-intersecting notification after setting display:none on target.");
  checkLastEntry(entries, 0, [8, 108, 8, 108, 8, 108, 8, 108, 0, window.innerWidth, 0, window.innerHeight, true]);
  target.style.display = "none";
}

function step1() {
  runTestCycle(step2, "Intersecting notification after removing display:none on target.");
  checkLastEntry(entries, 1, [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, false]);
  target.style.display = "";
}

function step2() {
  checkLastEntry(entries, 2, [8, 108, 8, 108, 8, 108, 8, 108, 0, window.innerWidth, 0, window.innerHeight, true]);
}
</script>
